+++ lib/common/index.ts
@@ -1,3 +1,7 @@
 
 export * from './prom.decorators';
+export * from './prom-counter.decorator';
+export * from './prom-gauge.decorator';
+export * from './prom-histogram.decorator';
+export * from './prom-summary.decorator';
 export { getMetricToken } from './prom.utils';