+++ lib/common/prom.decorators.ts
@@ -1,23 +1,24 @@
-import { Inject } from '@nestjs/common';
-import { getMetricToken, findOrCreateCounter } from './prom.utils';
-
-export const InjectCounterMetric = (name: string) => Inject(getMetricToken(`Counter`, name));
-export const InjectGaugeMetric = (name: string) => Inject(getMetricToken(`Gauge`, name));
-export const InjectHistogramMetric = (name: string) => Inject(getMetricToken(`Histogram`, name));
-export const InjectSummaryMetric = (name: string) => Inject(getMetricToken(`Summary`, name));
+import { findOrCreateCounter } from './prom.utils';
+import { IMetricArguments, CounterMetric } from '../interfaces';
 
 /**
  * Create and increment a counter when the method is called
  */
-export const PromMethodCounter = () => {
+export const PromMethodCounter = (params?: IMetricArguments) => {
     return (target: Object, propertyKey: string | symbol, descriptor: TypedPropertyDescriptor<(...args: any[]) => any>) => {
         const className = target.constructor.name;
-        const counterMetric = findOrCreateCounter({
-            name: `app_${className}_${propertyKey.toString()}_calls_total`,
-            help: `app_${className}#${propertyKey.toString()} called total`,
-        });
+        const name = `app_${className}_${propertyKey.toString()}_calls_total`;
+        const help = `app_${className}#${propertyKey.toString()} called total`;
+        let counterMetric: CounterMetric;
         const methodFunc = descriptor.value;
         descriptor.value = function (...args: any[]) {
+            if (!counterMetric) {
+                counterMetric = findOrCreateCounter({ 
+                    name,
+                    help,
+                    ...params,
+                });
+            }
             counterMetric.inc(1);
             return methodFunc.apply(this, args);
         };
@@ -29,15 +30,25 @@ export const PromMethodCounter = () => {
  *
  * @param ctor
  */
-export const PromInstanceCounter = <T extends { new(...args: any[]): {} }>(ctor: T) => {
-    const counterMetric = findOrCreateCounter({
-        name: `app_${ctor.name}_instances_total`,
-        help: `app_${ctor.name} object instances total`,
-    });
-    return class extends ctor {
-        constructor(...args) {
-            counterMetric.inc(1);
-            super(...args);
+export const PromInstanceCounter = (params?: IMetricArguments) => {
+    return <T extends { new(...args: any[]): {} }>(ctor: T) => {
+        const name = `app_${ctor.name}_instances_total`;
+        const help = `app_${ctor.name} object instances total`;
+        let counterMetric: CounterMetric;
+
+        return class extends ctor {
+            constructor(...args) {
+                if (!counterMetric) {
+                    counterMetric = findOrCreateCounter({
+                        name,
+                        help,
+                        ...params,
+                    });
+                }
+
+                counterMetric.inc(1);
+                super(...args);
+            }
         }
-    }
-};
+    };
+}