+++ lib/common/prom.utils.ts
@@ -1,5 +1,17 @@
-
 import * as client from 'prom-client';
+import { 
+  IMetricArguments, 
+  GenericMetric, 
+  CounterMetric, 
+  GaugeMetric, 
+  HistogramMetric, 
+  SummaryMetric, 
+  Registry,
+} from '../interfaces';
+
+type MetricType = 'Counter' | 'Gauge' | 'Histogram' | 'Summary';
+
+const registries = new Map<string, Registry>();
 
 export function getMetricToken(type: string, name: string) {
   return `${name}${type}`;
@@ -13,24 +25,70 @@ export function getOptionsName(name: string) {
   return `${name}PromOptions`;
 }
 
+export function getRegistry(name?: string) {
+  if (!name) {
+    return getDefaultRegistry();
+  }
+
+  if (registries.has(name) === false) {
+    const registry = new client.Registry();
+    registries.set(name, registry);
+  }
+
+  return registries.get(name);
+}
+
+export function getDefaultRegistry() {
+  return client.register;
+}
+
 export const findOrCreateMetric = ({
   name,
   type,
   help,
   labelNames,
+  registry,
 }: {
   name: string;
-  type: string;
+  type: MetricType;
   help?: string;
   labelNames?: string[];
-}): client.Metric<string> => {
-  let metric: client.Metric<string> = client.register.getSingleMetric(name);
+  registry?: Registry;
+}): GenericMetric => {
+
+  const register = registry ?? getDefaultRegistry();
+
+  let metric: GenericMetric = register.getSingleMetric(name);
   if (!metric) {
-    return new client.Counter({
-      name: name,
-      help: help || `${name} ${type}`,
-      labelNames,
-    });
+
+    switch (type) {
+      case "Gauge":
+        return new client.Gauge({
+          name: name,
+          help: help || `${name} ${type}`,
+          labelNames,
+        });
+      case "Histogram":
+        return new client.Histogram({
+          name: name,
+          help: help || `${name} ${type}`,
+          labelNames,
+        });
+      case "Summary":
+        return new client.Summary({
+          name: name,
+          help: help || `${name} ${type}`,
+          labelNames,
+        });
+      case "Counter":
+      default:
+        return new client.Counter({
+          name: name,
+          help: help || `${name} ${type}`,
+          labelNames,
+        });
+    }
+
   }
 
   if (metric instanceof client.Counter === false) {
@@ -40,73 +98,65 @@ export const findOrCreateMetric = ({
     });
   }
 
-  return metric;  
+  return metric;
 }
 
 export const findOrCreateCounter = ({
   name,
   help,
   labelNames,
-}: {
-  name: string;
-  help?: string;
-  labelNames?: string[];
-}): client.Counter<string> => {
+  registry,
+}: IMetricArguments): CounterMetric => {
   return findOrCreateMetric({
     name,
     help,
     type: `Counter`,
     labelNames,
-  }) as client.Counter<string>;
-};
+    registry,
+  }) as CounterMetric;
+}
 
 export const findOrCreateGauge = ({
   name,
   help,
   labelNames,
-}: {
-  name: string;
-  help?: string;
-  labelNames?: string[];
-}): client.Gauge<string> => {
+  registry,
+}: IMetricArguments): GaugeMetric => {
   return findOrCreateMetric({
     name,
     help,
     type: `Gauge`,
     labelNames,
-  }) as client.Gauge<string>;
-};
+    registry,
+  }) as GaugeMetric;
+}
 
 export const findOrCreateHistogram = ({
   name,
   help,
   labelNames,
-}: {
-  name: string;
-  help?: string;
-  labelNames?: string[];
-}): client.Histogram<string> => {
+  registry,
+}: IMetricArguments): HistogramMetric => {
   return findOrCreateMetric({
     name,
     help,
     type: `Histogram`,
     labelNames,
-  }) as client.Histogram<string>;
-};
+    registry,
+  }) as HistogramMetric;
+}
 
 export const findOrCreateSummary = ({
   name,
   help,
   labelNames,
-}: {
-  name: string;
-  help?: string;
-  labelNames?: string[];
-}): client.Summary<string> => {
+  registry,
+}: IMetricArguments): SummaryMetric => {
   return findOrCreateMetric({
     name,
     help,
     type: `Summary`,
     labelNames,
-  }) as client.Summary<string>;
-};
+    registry,
+  }) as SummaryMetric;
+}