+++ lib/index.ts
@@ -3,3 +3,5 @@ export * from './common';
 export * from './interfaces';
 export * from './prom.controller';
 export * from './prom.module';
+export * from './prom.service';
+export * from './middleware';