+++ lib/interfaces/index.ts
@@ -3,7 +3,9 @@ export * from './metric.type';
 export * from './prom-options.interface';
 
 // map class to expose
-export class CounterMetric<T extends string> extends PromClient.Counter<T> {}
-export class GaugeMetric<T extends string> extends PromClient.Gauge<T> {}
-export class HistogramMetric<T extends string> extends PromClient.Histogram<T> {}
-export class SummaryMetric<T extends string> extends PromClient.Summary<T> {}
+export type CounterMetric = PromClient.Counter<string>;
+export type GaugeMetric = PromClient.Gauge<string>;
+export type HistogramMetric = PromClient.Histogram<string>;
+export type SummaryMetric = PromClient.Summary<string>;
+export type GenericMetric = PromClient.Metric<string>;
+export type Registry = PromClient.Registry;