+++ lib/prom.constants.ts
@@ -2,5 +2,3 @@
 export const DEFAULT_PROM_REGISTRY = 'PromRegistry';
 export const DEFAULT_PROM_OPTIONS = 'PromOptions';
 export const PROM_REGISTRY_NAME = 'PromRegistryName';
-
-export const PROM_COUNTER_METRIC_AUTO_INC = 'PROM_COUNTER_METRIC_AUTO_INC';