+++ lib/prom.service.ts
@@ -1,21 +1,53 @@
 import { Injectable } from '@nestjs/common';
-import { findOrCreateCounter, findOrCreateGauge, findOrCreateHistogram } from './common/prom.utils';
+import {
+  findOrCreateCounter,
+  findOrCreateGauge,
+  findOrCreateHistogram,
+  findOrCreateSummary,
+  getDefaultRegistry,
+} from './common/prom.utils';
+import { IMetricArguments } from './interfaces';
 
 @Injectable()
 export class PromService {
+
+  getCounter(args: IMetricArguments) {
+    return findOrCreateCounter(args);
+  }
+
   getCounterMetric(name: string) {
-    return findOrCreateCounter({ name });
+    return this.getCounter({ name: name });
+  }
+
+  getGauge(args: IMetricArguments) {
+    return findOrCreateGauge(args);
   }
 
   getGaugeMetric(name: string) {
-    return findOrCreateGauge({ name });
+    return this.getGauge({ name: name });
+  }
+
+  getHistogram(args: IMetricArguments) {
+    return findOrCreateHistogram(args);
   }
 
   getHistogramMetric(name: string) {
-    return findOrCreateHistogram({ name });
+    return this.getHistogram({ name: name });
+  }
+
+  getSummary(args: IMetricArguments) {
+    return findOrCreateSummary(args);
   }
 
   getSummaryMetric(name: string) {
-    return findOrCreateHistogram({ name });
+    return this.getSummary({ name: name });
   }
+
+  /**
+   * Return the default registry
+   */
+  getDefaultRegistry() {
+    return getDefaultRegistry();
+  }
+
 }