+++ tests/src/app.service.ts
@@ -1,0 +1,15 @@
+import { Injectable } from '@nestjs/common';
+import { PromService } from '../../lib/prom.service';
+
+@Injectable()
+export class AppService {
+
+  constructor(
+    private readonly _promService: PromService,
+  ) {}
+
+  root(): string {
+    this._promService.getCounter({name: 'test'}).inc(1);
+    return 'Hello World!';
+  }
+}