+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## 1.0.0 - 2020-03-26
+
+### BREAKING CHANGE
+- Replace decorators
+  - `@InjectCounterMetric()` -> `@PromCounter()`
+  - `@InjectGaugeMetric()` -> `@PromGauge()`
+  - `@InjectHistogramMetric()` -> `@PromHistogram()`
+  - `@InjectSummaryMetric()` -> `@PromSummary()`
+- Remove `PromModule.forMetrics()` no need anymore
+- Rewrite internally how to manage
 
 ## 0.2.3 - 2020-07-26